# DriveSense-nest-api/docs/api.yaml
openapi: 3.0.0
info:
  title: DriveSense API
  description: API para gestionar restricciones de tráfico (pico y placa) en ciudades
  version: 1.0.0
  contact:
    name: DriveSense Team
servers:
  - url: http://localhost:4000
    description: Local Development Server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from login
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: John Doe
        email:
          type: string
          format: email
          example: john@example.com
        role:
          type: string
          enum: [admin, user]
          example: user
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    UpdateUserDto:
      type: object
      properties:
        name:
          type: string
          example: Jane Doe
        email:
          type: string
          format: email
          example: jane@example.com
        password:
          type: string
          minLength: 8
          example: NewSecurePass123
        role:
          type: string
          enum: [admin, user]
          example: admin
...
    Vehicle:
      type: object
      properties:
        id:
          type: integer
          example: 1
        licensePlate:
          type: string
          example: ABC-1234
        model:
          type: string
          example: Toyota Corolla
        color:
          type: string
          example: Blanco
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateVehicleDto:
      type: object
      required:
        - licensePlate
        - model
        - color
      properties:
        licensePlate:
          type: string
          example: ABC-1234
        model:
          type: string
          example: Toyota Corolla
        color:
          type: string
          example: Blanco
    UpdateVehicleDto:
      type: object
      properties:
        licensePlate:
          type: string
          example: XYZ-5678
        model:
          type: string
          example: Honda Civic
        color:
          type: string
          example: Negro

    # Rule Schemas
...
    UnauthorizedResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 401
        message:
          type: string
          example: Unauthorized
        error:
          type: string
          example: Unauthorized
    ForbiddenResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 403
        message:
          type: string
          example: Forbidden resource
        error:
          type: string
          example: Forbidden
    NotFoundResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 404
        message:
          type: string
          example: Resource not found
        error:
          type: string
          example: Not Found
    BadRequestResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 400
        message:
          type: array
          items:
            type: string
          example: ["Validation error"]
        error:
          type: string
          example: Bad Request
paths:
...
  /users:
    get:
      tags:
        - Users
      summary: Get all users
      description: List all users in the system (Admin only)
      security:
        - BearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
          description: Number of records per page
        - name: page
          in: query
          schema:
            type: integer
            default: 1
          description: Page number
      responses:
        200:
          description: Users retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedResponse'
        403:
          description: Forbidden - Admin role required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenResponse'
  /users/{id}:
    get:
      tags:
        - Users
      summary: Get user by ID
      description: Retrieve a specific user by their ID (Admin or the user themselves)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        200:
          description: User retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedResponse'
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenResponse'
        404:
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundResponse'

    put:
      tags:
        - Users
      summary: Update user
      description: Update user information (Admin or the user themselves)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserDto'
      responses:
        200:
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        400:
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedResponse'
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenResponse'
        404:
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundResponse'

    delete:
      tags:
        - Users
      summary: Delete user
      description: Remove a user from the system (Admin only)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        200:
          description: User deleted successfully
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedResponse'
        403:
          description: Forbidden - Admin role required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenResponse'
        404:
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundResponse'

  # Cities Endpoints
...
  /vehicles:
    post:
      tags:
        - Vehicles
      summary: Create a new vehicle
      description: Register a new vehicle (Users)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateVehicleDto'
      responses:
        201:
          description: Vehicle created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vehicle'
        400:
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedResponse'
        403:
          description: Forbidden - User role required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenResponse'

    get:
      tags:
        - Vehicles
      summary: Get all vehicles
      description: List all vehicles in the system (Admin only)
      security:
        - BearerAuth: []
      responses:
        200:
          description: Vehicles retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Vehicle'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedResponse'
        403:
          description: Forbidden - Admin role required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenResponse'
  /vehicles/{id}:
    get:
      tags:
        - Vehicles
      summary: Get vehicle by ID
      description: Retrieve a specific vehicle by its ID (Admin or Owner)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        200:
          description: Vehicle retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vehicle'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedResponse'
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenResponse'
        404:
          description: Vehicle not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundResponse'

    patch:
      tags:
        - Vehicles
      summary: Update vehicle
      description: Update vehicle information (Admin only)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateVehicleDto'
      responses:
        200:
          description: Vehicle updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vehicle'
        400:
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedResponse'
        403:
          description: Forbidden - Admin role required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenResponse'
        404:
          description: Vehicle not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundResponse'

    delete:
      tags:
        - Vehicles
      summary: Delete vehicle
      description: Remove a vehicle from the system (Admin only)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        200:
          description: Vehicle deleted successfully
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedResponse'
        403:
          description: Forbidden - Admin role required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenResponse'
        404:
          description: Vehicle not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundResponse'

  # Rules Endpoints
...