# DriveSense-nest-api/src/vehicles/entities/vehicle.entity.ts
import {
  Entity,
  PrimaryGeneratedColumn,
  Column,
  ManyToOne,
  CreateDateColumn,
  UpdateDateColumn,
  OneToMany,
} from 'typeorm';
import { User } from '../../users/entities/user.entity';
import { Log } from '../../logss/entities/log.entity';

export type VehicleType = 'car' | 'motorcycle';

@Entity('vehicles')
export class Vehicle {
  @PrimaryGeneratedColumn()
  id: number;

  @Column({ length: 20, unique: true })
  licensePlate: string;

  @Column({ length: 50 })
  brand: string;

  @Column({ length: 50 })
  model: string;

  @Column()
  year: number;

  @Column({ type: 'varchar', default: 'car' })
  type: VehicleType;

  @ManyToOne(() => User, user => user.id, { eager: true })
  user: User;

  @CreateDateColumn()
  createdAt: Date;

  @UpdateDateColumn()
  updatedAt: Date;

  @OneToMany(() => Log, (log) => log.vehicle)
  logs: Log[];
}